backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-jose[cryptography]
python-multipart
alembic
python-dotenv
passlib[bcrypt]
bcrypt
